@@ -14,10 +14,9 @@
     "@react-navigation/stack": "^5.2.9",
     "add": "^2.0.6",
     "expo": "^42.0.0",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
-    "react-native-elements": "^3.4.2",
     "react-native-gesture-handler": "~1.10.2",
     "react-native-reanimated": "~2.2.0",
     "react-native-safe-area-context": "3.2.0",
@@ -26,9 +25,6 @@
     "react-native-snap-carousel": "^3.8.0",
     "react-native-vector-icons": "^6.6.0",
     "react-native-web": "~0.13.12",
-    "react-navigation": "^4.3.5",
-    "react-navigation-drawer": "^2.4.9",
-    "react-navigation-stack": "^2.3.9",
     "react-redux": "^7.1.0",
     "yarn": "^1.22.17"
   },
