@@ -1,95 +1,65 @@
-import React from 'react';
-import {
-  FlatList,
-  Text,
-  View,
-  Image,
-  TouchableHighlight
-} from 'react-native';
-import styles from './styles';
-import { ListItem, SearchBar } from 'react-native-elements';
-import MenuImage from '../../components/MenuImage/MenuImage';
-import {
-  getCategoryName,
-  getRecipesByRecipeName,
-  getRecipesByCategoryName,
-  getRecipesByIngredientName
-} from '../../data/MockDataAPI';
+import React, { useEffect, useLayoutEffect, useState } from "react";
+import { FlatList, Text, View, Image, TouchableHighlight, Pressable } from "react-native";
+import styles from "./styles";
+import MenuImage from "../../components/MenuImage/MenuImage";
+import { getCategoryName, getRecipesByRecipeName, getRecipesByCategoryName, getRecipesByIngredientName } from "../../data/MockDataAPI";
+import { TextInput } from "react-native-gesture-handler";
 
-export default class SearchScreen extends React.Component {
-  
-  constructor(props) {
-    super(props);
-    const {navigation} = props
+export default function SearchScreen(props) {
+  const { navigation } = props;
+
+  const [value, setValue] = useState("");
+  const [data, setData] = useState([]);
+
+  useLayoutEffect(() => {
     navigation.setOptions({
-      headerLeft: ()=>
+      headerLeft: () => (
         <MenuImage
           onPress={() => {
             navigation.openDrawer();
           }}
         />
-      ,
-      headerTitle: ()=>
-        <SearchBar
-          containerStyle={{
-            backgroundColor: 'transparent',
-            borderBottomColor: 'transparent',
-            borderTopColor: 'transparent',
-            flex: 1
-          }}
-          inputContainerStyle={{
-            backgroundColor: '#EDEDED'
-          }}
-          inputStyle={{
-            backgroundColor: '#EDEDED',
-            borderRadius: 10,
-            color: 'black'
-          }}
-          searchIcond
-          clearIcon
-          //lightTheme
-          round
-          onChangeText={text => this.handleSearch(text)}
-          placeholder="Search"
-          value={this.getValue}
-        />,
-      headerRight: () => <View/>
-    })
-    this.state = {
-      value: '',
-      data: []
-    };
-  }
+      ),
+      headerTitle: () => (
+        <View style={styles.searchContainer}>
+          <Image style={styles.searchIcon} source={require("../../../assets/icons/search.png")} />
+          <TextInput
+            style={styles.searchInput}
+            onChangeText={handleSearch}
+            value={value}
+          />
+          <Pressable onPress={() => handleSearch("")}>
+          <Image style={styles.searchIcon} source={require("../../../assets/icons/close.png")} />
+          </Pressable>
+        </View>
+      ),
+      headerRight: () => <View />,
+    });
+  }, [value]);
 
-  handleSearch = text => {
+  useEffect(() => {}, [value]);
+
+  const handleSearch = (text) => {
+    setValue(text);
     var recipeArray1 = getRecipesByRecipeName(text);
     var recipeArray2 = getRecipesByCategoryName(text);
     var recipeArray3 = getRecipesByIngredientName(text);
     var aux = recipeArray1.concat(recipeArray2);
     var recipeArray = [...new Set(aux)];
-    if (text == '') {
-      this.setState({
-        value: text,
-        data: []
-      });
+
+    if (text == "") {
+      setData([]);
     } else {
-      this.setState({
-        value: text,
-        data: recipeArray
-      });
+      setData(recipeArray);
     }
   };
 
-  getValue = () => {
-    return this.state.value;
+  const onPressRecipe = (item) => {
+    navigation.navigate("Recipe", { item });
   };
 
-  onPressRecipe = item => {
-    this.props.navigation.navigate('Recipe', { item });
-  };
-
-  renderRecipes = ({ item }) => (
-    <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={() => this.onPressRecipe(item)}>
+  const renderRecipes = ({ item }) => (
+    <TouchableHighlight underlayColor="rgba(73,182,77,0.9)" onPress={() => onPressRecipe(item)}>
       <View style={styles.container}>
         <Image style={styles.photo} source={{ uri: item.photo_url }} />
         <Text style={styles.title}>{item.title}</Text>
@@ -98,18 +68,9 @@ export default class SearchScreen extends React.Component {
     </TouchableHighlight>
   );
 
-  render() {
-    return (
-      <View>
-        <FlatList
-          vertical
-          showsVerticalScrollIndicator={false}
-          numColumns={2}
-          data={this.state.data}
-          renderItem={this.renderRecipes}
-          keyExtractor={item => `${item.recipeId}`}
-        />
-      </View>
-    );
-  }
+  return (
+    <View>
+      <FlatList vertical showsVerticalScrollIndicator={false} numColumns={2} data={data} renderItem={renderRecipes} keyExtractor={(item) => `${item.recipeId}`} />
+    </View>
+  );
 }
