@@ -3,16 +3,14 @@ import { TouchableHighlight, Image, Text, View } from 'react-native';
 import PropTypes from 'prop-types';
 import styles from './styles';
 
-export default class ViewIngredientsButton extends React.Component {
-  render() {
+export default function ViewIngredientsButton (props) {
     return (
-      <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={this.props.onPress}>
+      <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={props.onPress}>
         <View style={styles.container}>
           <Text style={styles.text}>View Ingredients</Text>
         </View>
       </TouchableHighlight>
     );
-  }
 }
 
 ViewIngredientsButton.propTypes = {
