@@ -1,32 +1,32 @@
-import React from 'react';
-import { FlatList, Text, View, TouchableHighlight, Image } from 'react-native';
-import styles from './styles';
-import { recipes } from '../../data/dataArrays';
-import MenuImage from '../../components/MenuImage/MenuImage';
-import { getCategoryName } from '../../data/MockDataAPI';
+import React, { useLayoutEffect } from "react";
+import { FlatList, Text, View, TouchableHighlight, Image } from "react-native";
+import styles from "./styles";
+import { recipes } from "../../data/dataArrays";
+import MenuImage from "../../components/MenuImage/MenuImage";
+import { getCategoryName } from "../../data/MockDataAPI";
 
-export default class HomeScreen extends React.Component {
+export default function HomeScreen(props) {
+  const { navigation } = props;
 
-  constructor(props) {
-    super(props);
-    const {navigation} = props;
+  useLayoutEffect(() => {
     navigation.setOptions({
-      headerLeft: () => 
+      headerLeft: () => (
         <MenuImage
           onPress={() => {
             navigation.openDrawer();
           }}
-        />,
-      headerRight: () => <View/>
-    })
-  }
+        />
+      ),
+      headerRight: () => <View />,
+    });
+  }, []);
 
-  onPressRecipe = item => {
-    this.props.navigation.navigate('Recipe', { item });
+  const onPressRecipe = (item) => {
+    navigation.navigate("Recipe", { item });
   };
 
-  renderRecipes = ({ item }) => (
-    <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={() => this.onPressRecipe(item)}>
+  const renderRecipes = ({ item }) => (
+    <TouchableHighlight underlayColor="rgba(73,182,77,0.9)" onPress={() => onPressRecipe(item)}>
       <View style={styles.container}>
         <Image style={styles.photo} source={{ uri: item.photo_url }} />
         <Text style={styles.title}>{item.title}</Text>
@@ -35,18 +35,9 @@ export default class HomeScreen extends React.Component {
     </TouchableHighlight>
   );
 
-  render() {
-    return (
-      <View>
-        <FlatList
-          vertical
-          showsVerticalScrollIndicator={false}
-          numColumns={2}
-          data={recipes}
-          renderItem={this.renderRecipes}
-          keyExtractor={item => `${item.recipeId}`}
-        />
-      </View>
-    );
-  }
+  return (
+    <View>
+      <FlatList vertical showsVerticalScrollIndicator={false} numColumns={2} data={recipes} renderItem={renderRecipes} keyExtractor={(item) => `${item.recipeId}`} />
+    </View>
+  );
 }
