@@ -1,36 +1,28 @@
-import React from 'react';
-import {
-  FlatList,
-  ScrollView,
-  Text,
-  View,
-  Image,
-  TouchableHighlight
-} from 'react-native';
-import styles from './styles';
-import {
-  getIngredientUrl,
-  getRecipesByIngredient,
-  getCategoryName
-} from '../../data/MockDataAPI';
+import React, { useLayoutEffect } from "react";
+import { FlatList, ScrollView, Text, View, Image, TouchableHighlight } from "react-native";
+import styles from "./styles";
+import { getIngredientUrl, getRecipesByIngredient, getCategoryName } from "../../data/MockDataAPI";
 
-export default class IngredientScreen extends React.Component {
+export default function IngredientScreen(props) {
+  const { navigation, route } = props;
 
-  constructor(props) {
-    super(props);
-    const {navigation, route} = props
+  const ingredientId = route.params?.ingredient;
+  const ingredientUrl = getIngredientUrl(ingredientId);
+  const ingredientName = route.params?.name;
+
+  useLayoutEffect(() => {
     navigation.setOptions({
-      title: route.params?.name
-    })
-  }
+      title: route.params?.name,
+    });
+  }, []);
 
-  onPressRecipe = item => {
-    this.props.navigation.navigate('Recipe', { item });
+  const onPressRecipe = (item) => {
+    navigation.navigate("Recipe", { item });
   };
 
-  renderRecipes = ({ item }) => (
-    <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={() => this.onPressRecipe(item)}>
-      <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={() => this.onPressRecipe(item)}>
+  const renderRecipes = ({ item }) => (
+    <TouchableHighlight underlayColor="rgba(73,182,77,0.9)" onPress={() => onPressRecipe(item)}>
+      <TouchableHighlight underlayColor="rgba(73,182,77,0.9)" onPress={() => onPressRecipe(item)}>
         <View style={styles.container}>
           <Image style={styles.photo} source={{ uri: item.photo_url }} />
           <Text style={styles.title}>{item.title}</Text>
@@ -40,28 +32,15 @@ export default class IngredientScreen extends React.Component {
     </TouchableHighlight>
   );
 
-  render() {
-    const {  route } = this.props;
-    const ingredientId = route.params?.ingredient;
-    const ingredientUrl = getIngredientUrl(ingredientId);
-    const ingredientName = route.params?.name;
-    return (
-      <ScrollView style={styles.mainContainer}>
-        <View style={{ borderBottomWidth: 0.4, marginBottom: 10, borderBottomColor: 'grey' }}>
-          <Image style={styles.photoIngredient} source={{ uri: '' + ingredientUrl }} />
-        </View>
-        <Text style={styles.ingredientInfo}>Recipes with {ingredientName}:</Text>
-        <View>
-          <FlatList
-            vertical
-            showsVerticalScrollIndicator={false}
-            numColumns={2}
-            data={getRecipesByIngredient(ingredientId)}
-            renderItem={this.renderRecipes}
-            keyExtractor={item => `${item.recipeId}`}
-          />
-        </View>
-      </ScrollView>
-    );
-  }
+  return (
+    <ScrollView style={styles.mainContainer}>
+      <View style={{ borderBottomWidth: 0.4, marginBottom: 10, borderBottomColor: "grey" }}>
+        <Image style={styles.photoIngredient} source={{ uri: "" + ingredientUrl }} />
+      </View>
+      <Text style={styles.ingredientInfo}>Recipes with {ingredientName}:</Text>
+      <View>
+        <FlatList vertical showsVerticalScrollIndicator={false} numColumns={2} data={getRecipesByIngredient(ingredientId)} renderItem={renderRecipes} keyExtractor={(item) => `${item.recipeId}`} />
+      </View>
+    </ScrollView>
+  );
 }
