@@ -1,31 +1,27 @@
-import React from 'react';
-import {
-  FlatList,
-  Text,
-  View,
-  TouchableHighlight,
-  Image
-} from 'react-native';
-import styles from './styles';
-import { getRecipes, getCategoryName } from '../../data/MockDataAPI';
+import React, { useLayoutEffect } from "react";
+import { FlatList, Text, View, TouchableHighlight, Image } from "react-native";
+import styles from "./styles";
+import { getRecipes, getCategoryName } from "../../data/MockDataAPI";
 
-export default class RecipesListScreen extends React.Component {
+export default function RecipesListScreen(props) {
+  const { navigation, route } = props;
 
-  constructor(props) {
-    super(props);
-    const {navigation, route} = props
+  const item = route?.params?.category;
+  const recipesArray = getRecipes(item.id);
+
+  useLayoutEffect(() => {
     navigation.setOptions({
       title: route.params?.title,
-      headerRight: ()=> <View/>
-    })
-  }
+      headerRight: () => <View />,
+    });
+  }, []);
 
-  onPressRecipe = item => {
-    this.props.navigation.navigate('Recipe', { item });
+  const onPressRecipe = (item) => {
+    navigation.navigate("Recipe", { item });
   };
 
-  renderRecipes = ({ item }) => (
-    <TouchableHighlight underlayColor='rgba(73,182,77,0.9)' onPress={() => this.onPressRecipe(item)}>
+  const renderRecipes = ({ item }) => (
+    <TouchableHighlight underlayColor="rgba(73,182,77,0.9)" onPress={() => onPressRecipe(item)}>
       <View style={styles.container}>
         <Image style={styles.photo} source={{ uri: item.photo_url }} />
         <Text style={styles.title}>{item.title}</Text>
@@ -34,21 +30,9 @@ export default class RecipesListScreen extends React.Component {
     </TouchableHighlight>
   );
 
-  render() {
-    const {  route } = this.props;
-    const item = route?.params?.category;
-    const recipesArray = getRecipes(item.id);
-    return (
-      <View>
-        <FlatList
-          vertical
-          showsVerticalScrollIndicator={false}
-          numColumns={2}
-          data={recipesArray}
-          renderItem={this.renderRecipes}
-          keyExtractor={item => `${item.recipeId}`}
-        />
-      </View>
-    );
-  }
+  return (
+    <View>
+      <FlatList vertical showsVerticalScrollIndicator={false} numColumns={2} data={recipesArray} renderItem={renderRecipes} keyExtractor={(item) => `${item.recipeId}`} />
+    </View>
+  );
 }
